bpf/Makefile: Forbid implicit int conversions
This commit forbids implicit int conversions as well as conversions
between enumerations. A bunch of int conversions are made explicit as
well, but there shouldn't be any functional changes.

Enforcing this restriction at the compiler would have allowed us to
avoid 1cf3ef3f ("bpf: Fix invalid trace reason in bpf_host") if done
earlier. It allowed us to detect and fix another minor incorrect cast,
sent before this patch and merged as 108aa421 ("bpf: Fix implicit cast
for BPF TPROXY debug message").

Several implicit int conversions "fixed" here are inherited from u32 BPF
APIs (holding <32bit values for historic reasons). Other are due to our
use of 32-bit long slots for the metadata. Finally, others are due to
the get_prandom_u32 BPF helper returning a 32-bit integer, which can of
course be safely used as a smaller integer.

+++ bpf/lib/conntrack.h
@@ -1031,7 +1031,7 @@ static __always_inline bool
 ct_has_nodeport_egress_entry4(const void *map,
 			      struct ipv4_ct_tuple *ingress_tuple)
 {
-	int prev_flags = ingress_tuple->flags;
+	__u8 prev_flags = ingress_tuple->flags;
 	struct ct_entry *entry;
 
 	ingress_tuple->flags = TUPLE_F_OUT;
@@ -1048,7 +1048,7 @@ static __always_inline bool
 ct_has_nodeport_egress_entry6(const void *map,
 			      struct ipv6_ct_tuple *ingress_tuple)
 {
-	int prev_flags = ingress_tuple->flags;
+	__u8 prev_flags = ingress_tuple->flags;
 	struct ct_entry *entry;
 
 	ingress_tuple->flags = TUPLE_F_OUT;
